% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_analysis.R
\name{compute_chi}
\alias{compute_chi}
\title{Compute chi}
\usage{
compute_chi(st, starting_points, a0, mn, gisbase)
}
\arguments{
\item{st}{A SpatRaster object from `dem_process`}

\item{starting_points}{starting points (SpatVector) for the integration (outlets of basins)}

\item{a0}{reference accumulation in pixels}

\item{mn}{value of the m/n ratio}

\item{gisbase}{The directory path to GRASS binaries and libraries, containing bin and lib subdirectories among others}
}
\value{
a SpatRaster with chi values along streams
}
\description{
Compute chi
}
