% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_computation.R
\name{compute_curvature}
\alias{compute_curvature}
\title{Compute curvature over a raster}
\usage{
compute_curvature(rast, win, gisBase, planar = FALSE)
}
\arguments{
\item{rast}{a `RasterLayer` representing the raster (usually a Digital Elevation Model)}

\item{win}{an odd integer (>=3) specifying the size of the computation window (in pixels)}

\item{gisBase}{The directory path to GRASS binaries and libraries, containing bin and lib sub-directories among others}

\item{planar}{flag to compute planar curvature instead (default FALSE)}
}
\value{
a curvature raster
}
\description{
Fits quadratic surfaces at each pixel
}
\details{
The surface is parametrized as :
z = ax^2 + by^2 +cxy +dx + ey + f

curvature is C =  2a + 2b

the computation relies on GRASS GIS `r.param.scale` function
}
