% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_computation.R
\name{compute_slope}
\alias{compute_slope}
\title{Compute slope over a raster}
\usage{
compute_slope(rast, win, gisbase, central = TRUE)
}
\arguments{
\item{rast}{a `SpatRaster` representing the raster (usually a Digital Elevation Model)}

\item{win}{an odd integer (>=3) specifying the size of the computation window (in pixels)}

\item{gisbase}{The directory path to GRASS binaries and libraries, containing bin and lib sub-directories among others}

\item{central}{Constrain fitted surface through central window cell (default TRUE)}
}
\value{
a gradient `SpatRaster` (in m/m)
}
\description{
Fits quadratic surfaces at each pixel
}
\details{
The surfaces are parametrized as :
z = ax^2 + by^2 +cxy +dx + ey + f

Slope is calculated as S = (d^2 + e^2)^0.5

the computation relies on GRASS GIS `r.param.scale` function
}
