% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dem_cevennes}
\alias{dem_cevennes}
\title{Digital Elevation Model of the Cevennes area SE France}
\format{
A Raster Layer
\describe{
  \item{dimensions}{1568 rows and 2258 columns}
  \item{resolution}{0.0002777778}
  \item{extent}{E3.495139, E4.122361, N43.96764, N44.40319}
  \item{CRS}{+proj=longlat +datum=WGS84 +no_defs}
}
}
\source{
\url{https://opentopography.org/}
}
\usage{
dem_cevennes
}
\description{
The Shuttle Radar Topography Mission (SRTM) obtained elevation data on a near-global scale to generate the most complete high-resolution digital topographic database of Earth. SRTM consisted of a specially modified radar system that flew onboard the Space Shuttle Endeavour during an 11-day mission in February of 2000. SRTM is an international project spearheaded by the National Geospatial-Intelligence Agency (NGA) and the National Aeronautics and Space Administration (NASA).
}
\details{
NASA Shuttle Radar Topography Mission (SRTM)(2013). Shuttle Radar Topography Mission (SRTM) Global.

Distributed by OpenTopography.

 https://doi.org/10.5069/G9445JDF

 Accessed: 2020-12-04
}
\keyword{datasets}
