% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_next}
\alias{get_next}
\title{Get value of next pixel following flow direction.}
\usage{
get_next(rast, dir)
}
\arguments{
\item{rast}{Raster (RasterLayer)}

\item{dir}{Flow direction raster (RasterLayer)}
}
\value{

}
\description{
Compute for each pixel of a raster the value of the following pixel according to flow direction
}
