% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_srtm}
\alias{get_srtm}
\title{Dowload a STRM raster from srtm.csi.cgiar.org}
\usage{
get_srtm(sp, buf = 0.1)
}
\arguments{
\item{sp}{an object for which a spatial extent can be extracted}

\item{buf}{a buffering distance (in degrees) to expand the extent}
}
\value{

}
\description{
Compute for each pixel of a raster the value of the following pixel according to flow direction
}
