% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_processing.R
\name{get_streams}
\alias{get_streams}
\title{Extract portions of the stream network for river long-profile analysis}
\usage{
get_streams(net, id, mode = "up")
}
\arguments{
\item{net}{stream network from get_network}

\item{id}{id of the starting stream segment (integer)}

\item{mode}{type of extraction : upstream from id (`up`, default), upstream including all tributaries (`up_all`) or downstream from id (`down`).}
}
\value{
If `mode` is `up` (default) returns a vector with the upstream ids.
If `mode` is `up_all` returns a dataframe with the first line corresponding to the main stream, and following lines to tributaries.
The column `ids` contains the ids of the stream segments, `length` the total length of the stream and `n` the number of segments.
If `mode` is `down` returns a vector with the downstream ids.
}
\description{
Get ids for all streams segments moving up (default) or down from a starting stream segment (identified with `id`).
Follow the largest accumulation tributary at confluences when moving upstream.
}
