% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{getA}
\alias{getA}
\title{Perron 2011}
\usage{
getA(z, f, F, nx, dt)
}
\arguments{
\item{z}{elevation vector (m)}

\item{f}{f}

\item{F}{F}

\item{nx}{number of space increment}

\item{dt}{time step (a)}
}
\value{

}
\description{
Perron 2011
}
