% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{getf}
\alias{getf}
\title{Perron 2011}
\usage{
getf(z, K, prps, Sc, nx, dx, U)
}
\arguments{
\item{z}{elevation vector (m)}

\item{K}{hillslope diffusion coefficient (m2/a)}

\item{prps}{rock-to-regolith density ratio (-)}

\item{Sc}{critical hillslope gradient (m/m)}

\item{nx}{number of space increment}

\item{dx}{horizontal distance increment (m)}

\item{U}{relative baselevel fall rate (m/a)}
}
\value{

}
\description{
Perron 2011
}
