% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{hillslope_Es}
\alias{hillslope_Es}
\title{Compute hillslope non-dimensional Erosion rate E*}
\usage{
hillslope_Es(U, K, L, beta, Sc)
}
\arguments{
\item{U}{relative baselevel fall rate (m/a)}

\item{K}{hillslope diffusion coefficient (m2/a)}

\item{L}{hillslope length (m)}

\item{beta}{rock-to-regolith density ratio (-)}

\item{Sc}{critical hillslope gradient (m/m)}
}
\value{
Value of E*, which can be a vector
}
\description{
Compute hillslope non-dimensional Erosion Rate E*  according to \insertCite{roering2007functional;textual}{geomorphology}
}
\references{
\insertAllCited{}
}
