% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{hillslope_profile}
\alias{hillslope_profile}
\title{Hillslope steady state profiles}
\usage{
hillslope_profile(x, E, K, L, beta, Sc = NULL)
}
\arguments{
\item{x}{Horizontal position along the hillslope (x=0 at hilltop, m)}

\item{E}{Denudation rate (m/a)}

\item{K}{Diffusion coefficient (m2/a)}

\item{L}{hillslope length (m)}

\item{beta}{rock to teglith density ratio (dimensionless)}

\item{Sc}{Critical hillslope gradient (optional, m/m)}
}
\value{
Hillslope elevation (relative to base of hillslope) at the position specified in `x`.
}
\description{
Compute hillslope topography steady state profiles, in the case of linear and nonlinear sediment flux
}
\details{
If the threshold hillslope gradient `Sc` is specified the profile is computed in the non-linear case, otherwise linear.
}
\examples{
x = seq(0,100,by=1)
L =  max(x) # hillslope length
E = 100/1e6 # erosion rate (m/Ma -> m/a)
K = 0.01 # diffusion coefficient (m2/a)
beta = 2
Sc = 0.8 # (m/m)
znl = hillslope_profile(x,E,K,L,beta,Sc)
plot(x,znl,type="l",lwd=3,
xlab="Distance from hilltop (m)",ylab="Elevation (m)")
zl = hillslope_profile(x,E,K,L,beta)
lines(x,zl,col="blue")
}
