% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{hillslope_psi}
\alias{hillslope_psi}
\title{Ratio of the nonlinear to linear components of the sediment flux}
\usage{
hillslope_psi(U, K, L, beta, Sc)
}
\arguments{
\item{U}{relative baselevel fall rate (m/a)}

\item{K}{hillslope diffusion coefficient (m2/a)}

\item{L}{hillslope length (m)}

\item{beta}{rock-to-regolith density ratio (-)}

\item{Sc}{critical hillslope gradient (m/m)}
}
\value{
Value of Psi, which can be a vector
}
\description{
Compute ratio of the nonlinear to linear components of sediment flux at the base of the hillslope according to \insertCite{roering2001hillslope;textual}{geomorphology}
}
\references{
\insertAllCited{}
}
