% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{hillslope_tau}
\alias{hillslope_tau}
\title{Hillslope adjustement time}
\usage{
hillslope_tau(U, K, L, beta, Sc = NULL, A = 0.4052847, B = 1.74)
}
\arguments{
\item{U}{relative baselevel fall rate (m/a)}

\item{K}{hillslope diffusion coefficient (m2/a)}

\item{L}{hillslope length (m)}

\item{beta}{rock-to-regolith density ratio (-)}

\item{Sc}{critical hillslope gradient (m/m), optional if not specified the linear adjustement timescale is returned}

\item{A}{constant from \insertCite{roering2001hillslope;textual}{geomorphology}, optional (default 0.405)}

\item{B}{constant from \insertCite{roering2001hillslope;textual}{geomorphology}, optional (default 1.74)}
}
\value{

}
\description{
Compute exponential equilibrium adjustment timescale for sediment flux or hillslope morphology according to \insertCite{roering2001hillslope;textual}{geomorphology}
}
\references{
\insertAllCited{}
}
