% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_incision.R
\name{I_instantaneous}
\alias{I_instantaneous}
\title{Compute instantaneous incision}
\usage{
I_instantaneous(Qs, ks, Psi, K, n, gamma)
}
\arguments{
\item{Qs}{Normalized discharge (to mean annual discharge)}

\item{ks}{Channel steepness index}

\item{Psi}{Incision threshold}

\item{K}{erodibility}

\item{n}{slope exponent of the stream power law}

\item{gamma}{discharge exponent}
}
\value{
Instantaneous incision value (m/s)
}
\description{
See equation 8 of Dibiase and Whipple (2011)
}
