% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_incision.R
\name{I_long_term}
\alias{I_long_term}
\title{Compute long-term channel incision according to a stream power-based instantaneous incision law and an inverse gamma distribution for discharge}
\usage{
I_long_term(k, ks, Psi, K, n, gamma)
}
\arguments{
\item{k}{Variability parameter of the inverse gamma distribution for discharge}

\item{ks}{Channel steepness index}

\item{Psi}{Incision threshold}

\item{K}{erodibility}

\item{n}{slope exponent of the stream power law}

\item{gamma}{discharge exponent}
}
\value{
Long-term incision value (m/s)
}
\description{
This function is actually solving equation 10 from Dibiase and Whipple (2011)
}
