% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_incision.R
\name{Inv_Gamma_dist}
\alias{Inv_Gamma_dist}
\title{Compute probability density of discharge according to an inverse gamma law}
\usage{
Inv_Gamma_dist(k, Q, Qb)
}
\arguments{
\item{k}{Variability parameter}

\item{Q}{Discharge}

\item{Qb}{Mean annual discharge}
}
\value{
Probability density values for discharge
}
\description{
Q and Qb must have the same units
}
\details{
See equation 11 of Dibiase and Whipple (2011)
}
