% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outlets_cevennes}
\alias{outlets_cevennes}
\title{Basins outlets positions in the Cevennes area SE France}
\format{
A SpatialPointsDataFrame
\describe{
  \item{features}{5}
  \item{variables}{2}
  \item{extent}{565391.4, 584650.7, 4881077, 4892994}
  \item{CRS}{+proj=utm +zone=31 +datum=WGS84 +units=m +no_defs}
}
}
\usage{
outlets_cevennes
}
\description{
Basins outlets positions in the Cevennes area SE France
}
\keyword{datasets}
