% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillslope_modelling.R
\name{perron_nlhs}
\alias{perron_nlhs}
\title{Nonlinear hillslope evolution solver}
\usage{
perron_nlhs(z, K, beta, Sc, U, dx, dt)
}
\arguments{
\item{z}{elevation vector (m)}

\item{K}{hillslope diffusion coefficient (m2/a)}

\item{beta}{rock-to-regolith density ratio (-)}

\item{Sc}{critical hillslope gradient (m/m)}

\item{U}{relative baselevel fall rate (m/a)}

\item{dx}{horizontal distance increment (m)}

\item{dt}{time step (a)}
}
\value{
Elevation vector at the end of time step
}
\description{
Solve nonlinear hillslope evolution using \insertCite{perron2011numerical;textual}{geomorphology} implicit formulation over one time step
}
\references{
\insertAllCited{}
}
