% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profile_cevennes}
\alias{profile_cevennes}
\title{Profile line across the Cevennes area SE France}
\format{
A SpatialLinesDataFrame
\describe{
  \item{features}{1}
  \item{variables}{0}
  \item{extent}{552118.7, 587431.2, 4880279, 4914009}
  \item{CRS}{+proj=utm +zone=31 +datum=WGS84 +units=m +no_defs}
}
}
\usage{
profile_cevennes
}
\description{
Profile line across the Cevennes area SE France
}
\keyword{datasets}
