% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_profile_analysis.R
\name{profile_gaussian}
\alias{profile_gaussian}
\title{profile_gaussian}
\usage{
profile_gaussian(distance, elevation, span, sigma)
}
\arguments{
\item{distance}{distance along channel network (m)}

\item{elevation}{elevation along channel network (m)}

\item{span}{span parameter for the loess fit used to detrend the profile}

\item{sigma}{standard deviation of the Gaussian window (m)}
}
\value{

}
\description{
Perform Gaussian filtering of river profile
}
