% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_profile_analysis.R
\name{profile_gradient}
\alias{profile_gradient}
\title{profile_gradient}
\usage{
profile_gradient(distance, elevation, window)
}
\arguments{
\item{distance}{distance along channel network (m)}

\item{elevation}{elevation along channel network (m)}

\item{window}{size of the window for the computation of the gradient}
}
\value{

}
\description{
Compute river profile gradient over a given window
}
