% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/river_profile_analysis.R
\name{profile_smooth}
\alias{profile_smooth}
\title{Smooth channel profile by removing artifacts}
\usage{
profile_smooth(distance, elevation, span, radius)
}
\arguments{
\item{distance}{distance along channel network (m)}

\item{elevation}{elevation along channel network (m)}

\item{span}{span parameter for the loess fit used to detrend the profile (fraction of the total points used for the local fits)}

\item{radius}{radius of the rolling circle used to smooth the profile (m)}
}
\value{

}
\description{
Smooth channel profile by removing artifacts
}
