% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{project_points}
\alias{project_points}
\title{Project points along a section.}
\usage{
project_points(x, y, X1, Y1, X2, Y2, width = NULL, inc = NULL, value = NULL)
}
\arguments{
\item{x}{X coordinates of the points to be projected (vector)}

\item{y}{Y coordinates of the points to be projected (vector)}

\item{X1}{X coordinate of starting point defining the projection line}

\item{Y1}{Y coordinate of starting point defining the projection line}

\item{X2}{X coordinate of end point defining the projection line}

\item{Y2}{Y coordinate of end point defining the projection line}

\item{width}{half width of the projection swath (optional)}

\item{inc}{number of bins along the profile for the stats (optional)}

\item{value}{parameter for which the statistics should be returned (optional, vector of same length and order as `x` and `y`)}
}
\value{
A dataframe with the distance along and to the sections of the input points. If `width`, `inc` and `value` are supplied a dataframe containing the binned (`inc` bins) statistics of `value` along the section is returned instead
}
\description{
Return a dataframe with two columns (same row order as the input)
`distance_along` : distance along  the profile of the projections
`distance_to` : perpendicular distance to the profile of the points
All input should be in cartesian coordinates with the same units
}
