% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_raster}
\alias{project_raster}
\title{Raster projection utility}
\usage{
project_raster(rast, epsg, resolution, method = "bilinear", trim = FALSE)
}
\arguments{
\item{rast}{input raster (`SpatRaster`)}

\item{epsg}{the EPSG code to use for projecting the raster}

\item{resolution}{the resolution of the projected raster (pixel size in m)}

\item{method}{the projection method (default `bilinear`, see `project` from `terra` package for other available options)}

\item{trim}{flag (default `FALSE`) to trim the periphery of the rast after projection and remove all NA pixels (use with caution as it can shrink substantially the output raster)}
}
\value{
a projected raster
}
\description{
Project a raster according to a specified CRS and resolution.
It will adjust the output extent according to the resolution.
}
