% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_incision.R
\name{Qcritical}
\alias{Qcritical}
\title{Compute critical discharge (normalized) for the onset of incision}
\usage{
Qcritical(ks, Psi, K, n, gamma)
}
\arguments{
\item{ks}{Channel steepness index}

\item{Psi}{Incision threshold}

\item{K}{erodibility}

\item{n}{slope exponent of the stream power law}

\item{gamma}{discharge exponent}
}
\value{
Normalized critical discharge
}
\description{
Determined by setting I=0 in equation 8 (instantaneous discharge) of Dibiase and Whipple (2011)
}
