% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers_cevennes}
\alias{rivers_cevennes}
\title{Along stream data for 5 rivers in the Cevennes area SE France}
\format{
A dataframe
\describe{
  \item{features}{1}
  \item{variables}{10}
  \item{extent}{565391.4, 584650.7, 4881077, 4892994}
  \item{CRS}{+proj=utm +zone=31 +datum=WGS84 +units=m +no_defs}
}
}
\usage{
rivers_cevennes
}
\description{
Along stream data for 5 rivers in the Cevennes area SE France
}
\keyword{datasets}
