% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start_grass}
\alias{start_grass}
\title{Initiate a grass throwaway session with proper region definition}
\usage{
start_grass(rast, name, gisBase)
}
\arguments{
\item{rast}{Raster to import (SpatRaster)}

\item{name}{for the raster in the location}

\item{gisBase}{The directory path to GRASS binaries and libraries, containing bin and lib subdirectories among others}
}
\value{

}
\description{
Initiate a grass throwaway session with proper region definition
}
