% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{streams2basins}
\alias{streams2basins}
\title{Compute continuous maps for variable defined along the stream network}
\usage{
streams2basins(st, rst, FUN = "mean")
}
\arguments{
\item{st}{A SpatRaster object from `dem_process`}

\item{rst}{A SpatRaster with the variable of interest (defined along streams), which will be transferred to the corresponding contributing areas.}

\item{FUN}{The function used to summarize the values for each stream segment (character string : mean, min, max, median)}
}
\value{

}
\description{
Transfer statistic values from stream segments to the corresponding contributing basins.
It allows to provide a continuous representation of variables defined along the stream network such as chi, for visualization purpose.
}
