% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_data.R
\name{swath_profile}
\alias{swath_profile}
\title{Project raster data along a swath profile}
\usage{
swath_profile(rast, X1, Y1, X2, Y2, width, inc)
}
\arguments{
\item{rast}{input raster}

\item{X1}{X coordinate of starting point defining the projection line}

\item{Y1}{Y coordinate of starting point defining the projection line}

\item{X2}{X coordinate of end point defining the projection line}

\item{Y2}{Y coordinate of end point defining the projection line}

\item{width}{one-side width of the swath profile}

\item{inc}{number of bins along the profile for the stats}
}
\value{
a data frame with statistics computed for each increment along the profile
}
\description{
The profile is defined by its end points and its width.
The results are returned as summary statistics for equally spaced bins along the profile.
}
