% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_computation.R
\name{compute_zonal_stats}
\alias{compute_zonal_stats}
\title{Compute zonal statistics from a raster layer}
\usage{
compute_zonal_stats(vect, rast, name, gisbase)
}
\arguments{
\item{vect}{the vector layer used for computation}

\item{rast}{a `RasterLayer` representing the raster}

\item{name}{a prefix for the statistics columns}

\item{gisBase}{The directory path to GRASS binaries and libraries, containing bin and lib sub-directories among others}
}
\value{
a vector with attribute table filled with statistics from the raster layer
}
\description{
This function is a frontend for GRASS module `v.rast.stats`
}
