% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_processing.R
\name{get_basins}
\alias{get_basins}
\title{Compute basins corresponding to outlets}
\usage{
get_basins(st, outlets, gisbase)
}
\arguments{
\item{st}{A SpatRaster object from `dem_process`}

\item{outlets}{Outlets points (SpatVector)}

\item{gisbase}{The directory path to GRASS binaries and libraries, containing bin and lib sub-directories among others}
}
\value{
A SpatVector with the corresponding basins. The attribute table of the outlets vector is transferred.
}
\description{
Follow the largest accumulation tributary at confluences when moving upstream
This function relies on the following GRASS GIS modules :
r.stream.basins (add-on to be installed)
}
