% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_processing.R
\name{get_outlets}
\alias{get_outlets}
\title{Extract basins outlets locations with various approaches. These various options are mutually exclusives.}
\usage{
get_outlets(st, strahler = NA, large = NA, elevation = NA, gisbase = NA)
}
\arguments{
\item{st}{A SpatRaster multi-layers object from `dem_process`}

\item{strahler}{The Strahler order of the outlets to be extracted}

\item{large}{The minimum size (in pixels) for the extraction of the largest possible set of basins}

\item{elevation}{The elevation of the outlets}

\item{gisbase}{In the case of elevation option, the directory path to GRASS binaries and libraries, containing bin and lib subdirectories among others}
}
\value{
a SpatVector with identified outlets
}
\description{
option `strahler` extract all the basins of the given strahler order
}
\details{
option `large` extract the largest possible basins from the DEM.
The indicated minimum size should be larger than the accumulation threshold used to delineate the network.

option `elevation` extract all outlets at a given elevation
}
