% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_na}
\alias{trim_na}
\title{Trim a raster to remove NA values on the periphery}
\usage{
trim_na(rst0)
}
\arguments{
\item{rast}{input raster}
}
\value{
a trimmed raster
}
\description{
The trim function from the raster package ony remove cols and rows full of NA
Useful to get a nicely formatted raster grid after reprojection
}
